top N calculation

+++ techbot/src/main/kotlin/com/firelib/techbot/command/SettingsCommand.kt
@@ -39,7 +39,7 @@ class SettingsCommand : TextCommand{
         })
 
         bot.sendMessage(
-            chatId = ChatId.fromId(userId.toLong()),
+            chatId = ChatId.fromId(userId),
             text = header + txt,
             parseMode = ParseMode.MARKDOWN_V2
         )
